moved components/keys to its own top level package
