+++ storage/kv/kv_data_storage.go
@@ -14,12 +14,15 @@
 package kv
 
 import (
+	"fmt"
 	"math"
+	"os"
 	"sync"
 	"sync/atomic"
 
 	"github.com/fagongzi/util/format"
-	"github.com/matrixorigin/matrixcube/components/keys"
+
+	"github.com/matrixorigin/matrixcube/keys"
 	"github.com/matrixorigin/matrixcube/pb/meta"
 	"github.com/matrixorigin/matrixcube/storage"
 	"github.com/matrixorigin/matrixcube/storage/stats"
@@ -135,6 +138,7 @@ func (kv *kvDataStorage) SaveShardMetadata(metadatas []storage.ShardMetadata) er
 	seen := make(map[uint64]struct{})
 	kv.mu.Lock()
 	for _, m := range metadatas {
+		fmt.Fprintf(os.Stderr, "shard: %d, index: %d\n", m.ShardID, m.LogIndex)
 		wb.Set(keys.GetMetadataKey(m.ShardID, m.LogIndex, nil), m.Metadata)
 		wb.Set(keys.GetAppliedIndexKey(m.ShardID, nil), format.Uint64ToBytes(m.LogIndex))
 		kv.mu.lastAppliedIndexes[m.ShardID] = m.LogIndex
@@ -161,7 +165,9 @@ func (kv *kvDataStorage) GetInitialStates() ([]storage.ShardMetadata, error) {
 	var shards []uint64
 	var lastApplied []uint64
 	// find out all shards and their last applied indexes
+	fmt.Printf(" ===> going to call scan!!!\n")
 	if err := kv.base.Scan(min, max, func(key, value []byte) (bool, error) {
+		fmt.Printf(" ===> got a key!!!\n")
 		if keys.IsAppliedIndexKey(key) {
 			shardID, err := keys.GetShardIDFromAppliedIndexKey(key)
 			if err != nil {
@@ -174,6 +180,7 @@ func (kv *kvDataStorage) GetInitialStates() ([]storage.ShardMetadata, error) {
 	}, false); err != nil {
 		return nil, err
 	}
+	fmt.Printf("shards len: %d\n", len(shards))
 	// update the persistentAppliedIndexes
 	kv.mu.Lock()
 	for idx, appliedIndex := range lastApplied {
